Move make_slow_nested_set to basis as classmethod
